requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
bs4==0.0.1
click==6.7
decorator==4.0.11
Flask==0.12
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.9.4
MarkupSafe==0.23
packaging==16.8
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
update-checker==0.16
Werkzeug==0.11.15
lxml==3.6.0
urllib3==1.21.1
APScheduler==3.0.0
sqlalchemy
flask_sqlalchemy
python-dateutil